Create Grammar.txt
These are mostly a few notes and ideas about what I might want to include in Language. Everything is subject to change, however, examples without a question mark are less likely to be changed than ones with a question mark. I know it's unorganized. In the future, I would like to split the sections into their own files and then expand upon them.